+++ src/components/SlotMachine.tsx
@@ -4,20 +4,7 @@ import Box from "@mui/material/Box";
 import Button from "@mui/material/Button";
 import {BET_COINS, INITIAL_COINS, SYMBOLS} from "./const"; // SymbolTileのインポート
 import {SymbolTile} from "./type";
-
-const spinReels = (
-  currentReels: SymbolTile[],
-  symbols: SymbolTile[],
-  betCoins: number
-): {newReels: SymbolTile[]; win: boolean; points: number} => {
-  const newReels = currentReels.map(
-    () => symbols[Math.floor(Math.random() * symbols.length)]
-  );
-  const win = new Set(newReels.map((tile) => tile.id)).size === 1;
-  const points = win ? newReels[0].point : 0;
-
-  return {newReels, win, points};
-};
+import {spinReels} from "./func";
 
 export const SlotMachine: React.FC = () => {
   const [coins, setCoins] = useState(INITIAL_COINS);
